pyproject.toml
[project]
name = "django_ninja_pdm_temp"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "yosong", email = "2404559603@qq.com"},
]
dependencies = ["django==4.2", "python-dotenv>=1.1.0", "django-ninja>=1.4.1", "django-redis>=5.4.0", "pyjwt>=2.10.1"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]

[tool.ruff]
ignore = [
"E402"
]
select = [
    "I",   # isort
    "E",   # pycodestyle
    "F",   # pyflakes
    "UP", # pyupgrade
]
line-length = 130
exclude = ["migrations"]

[tool.mypy]
exclude = 'migrations'


[[tool.mypy.overrides]]
module = "apps.user.*"
follow_imports = "skip"
ignore_errors = true
